refactor: new step button
