+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/NewStepButton/index.ts
@@ -1,0 +1,1 @@
+export { NewStepButton } from './NewStepButton'
